ww_test/requirements.txt
airtest==1.3.0.1
BeautifulReport==0.1.3
pyautogui==0.9.53
pytesseract==0.3.10
Pillow==9.4.0
pyyaml==6.0.1
xlwt==1.3.0
xlrd==2.0.1
pandas==2.1.0
ddt==1.7.0
pocoui==1.0.93
